models: add support for reactions
This commit adds support for reactions in the IRC client, and routes
them into the model. This commit does not yet allow for viewing
reactions, but it should store them correctly in the model for doing so.

This commit also fixes a bug where TAGMSGs (and other non-
PRIVMSG/NOTICE) commands were stored in the model. In the future maybe
this can be prevented by using a type-safe description of irc messages.

We allow PRIVMSG-based reacts, and assume that they are a fallback for
clients which do not support reactions. Thus we do not show the PRIVMSG,
only the reaction.