client_controller: fix bogus sync mode
When receiving RPL_ENDOFMOTD, some CAP REQs might still be in
flight, especially during first connection.

The cap being available should be enough to guarantee that we'll
at some point be able to use the cap (even if not right away).

Ideally we'd be listening to cap changes but I don't want to spam
workmanager/push/foreground service calls.