Add custom user agent
Some sites (like https://0x0.st) prefer custom client user agents, so
that they can identify where traffic is coming from (preferably with a
comment leading upstream). This commit changes the user agent for
goguma.

It is not enough to override `HttpClient` use around the codebase, nor
set headers for the image previews because how flutter is structures.
Neither of these are able to set all user agents sent in the app.
Overriding the http client implementation does this, so it is the route
that was taken.