Save composer text in database
This modifies ComposerState._replyTo to only have the msgid, since we
don't want to store the whole message in database, and it's only used
for that.

The draft is saved when the user leaves the buffer screen (either by
pressing back, going in the background or by quitting the app entirely).
I don't know if the model should be updated on every key press or if
it's too much.

In the model, draft is a class instead of two fields so that
notifyListeners() is only called once on draft update.

In the BufferEntry, draft is two fields to reflect the database schema.

Co-authored-by: Simon Ser <contact@emersion.fr>