Allow self-replies
This is useful when you want to link to a message you said earlier.
Other platforms like Discord, Signal and Whatsapp allow this.

Since it doesn't make sense on a server that doesn't relay the
`+reply` tag, this feature checks for CLIENTTAGDENY and shows an error
message to the user if it's not available.