Color the dot before actions in the sender's nick color, make the nick color of actions the same as the text color.
With the changes to message bubble colors the look of actions has been
changed too. The dot before the action was very subtle and hard to
notice. This commit changes the dot color to the sender's nick color,
back how it was before the message bubble colors were changed. Also the
nick color of actions is changed to the text color.